+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Move FHIR Packages to dktk-exporter
 - Refactoring converter manager
 - Rename page-counter as page
+- Rename FHIR_QUERY to FHIR_PATH and CQL_QUERY to CQL
 
 ### Fixed
 - JSON and XML filter